+++ include/string.h
@@ -4,8 +4,8 @@
 #ifndef STRING_H
 #define STRING_H
 
-void* memcpy(void*, const void*, uint32_t);
-void* memset(void*, uint8_t, uint32_t);
+void* memcpy(void*, const void*, size_t);
+void* memset(void*, uint8_t, size_t);
 int32_t strcmp(const char*, const char*);
 size_t strlen(const char*);
